% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeIDGF.R
\name{computeIDGF}
\alias{computeIDGF}
\title{Calcul de l'IDGF}
\usage{
computeIDGF(IDGFdata)
}
\arguments{
\item{IDGFdata}{Données importées par `importIDGF()`}
}
\value{
une liste
}
\description{
Calcul de l'IDGF
}
