% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radar_IDGF.R
\name{Diagnostic_IDGF}
\alias{Diagnostic_IDGF}
\title{Sorties graphiques de l'IDGF}
\usage{
Diagnostic_IDGF(result_IDGF)
}
\arguments{
\item{result_IDGF}{Les résultats issus de la fonction `IDGF()`}
}
\value{
Graphiques d'appui au diagnostic pour chaque relevé présent dans le tableau initial
}
\description{
Produit des diagrammes d'appui au diagnostic basés sur les résultats d'un IDGF
}
\examples{
res.IDGF <- IDGF(taxa.GF)
Diagnostic_IDGF(res.IDGF)
}
