% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportIDGF.R
\name{exportIDGF}
\alias{exportIDGF}
\title{Export des résultats et graphiques issus de l'IDGF}
\usage{
exportIDGF(IDGFres, outdir = paste0("RES_", Sys.Date()))
}
\arguments{
\item{IDGFres}{Résultats IDGF issus de `computeIDGF()` ou `radarIDGF()`}

\item{outdir}{Dossier de destination des fichiers, à définir par l'utilisateur}
}
\value{
Les résultats sont exportés dans le dossier défini par l'utilisateur
}
\description{
Cette fonction permet de sauvegarder les résultats issus de l'applicatif IDGF sous forme de tableau (.csv) et de graphiques (.png) dans un répertoire défini par l'utilisateur.
}
\examples{
library(IDGF)
data <- system.file("input_test.xlsx", package = "IDGF")
IDGFdata <- importIDGF(data)
IDGFres <- computeIDGF(IDGFdata)
IDGFresrad <- radarIDGF(IDGFres)
exportIDGF(IDGFresrad, outdir = "sorties")
}
