% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDGF.R
\name{IDGF}
\alias{IDGF}
\title{Indice Diatomique de Guyane Française}
\usage{
IDGF(df, lang = "FR")
}
\arguments{
\item{df}{Le tableau contenant la liste floristique, composé de 4 colonnes (dans cet ordre) : id_releve, cd_taxon, abondance, her (1 pour la plaine littorale ou 2 pour le bouclier Guyanais).}

\item{lang}{Argument acceptant deux valeurs : "FR" pour obtenir des résultats en français, "ENG" pour obtenir des résultats en anglais}
}
\value{
Un tableau détaillant les valeurs de chaque métrique individuelle (exprimée en EQR), une indication sur la robustesse sur l'indice ainsi que la proportion (et l'identité) des taxons inconnus et halins
}
\description{
Calcule l'indice IDGF à partir d'un ou plusieurs inventaires diatomiques
}
\examples{
head(taxa.GF)taxa.GF
IDGF(taxa.GF, lang = "FR")
}
