% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radarIDGF.R
\name{radarIDGF}
\alias{radarIDGF}
\title{Production de diagrammes de diagnostic issus de l'IDGF}
\usage{
radarIDGF(IDGFres)
}
\arguments{
\item{IDGFres}{sortie de la fonction `computeIDGF()`}
}
\value{
Renvoie le tableau issu de la fonction `computeIDGF()` en y rajoutant une colonne contenant les graphiques
}
\description{
Cette fonction s'appuie sur les résultats issus de `computeIDGF()` et y rajoute un diagramme d'appui au diagnostic pour chaque opération de contrôle.
}
\examples{
library(IDGF)
data <- system.file("input_test.xlsx", package = "IDGF")
IDGFdata <- importIDGF(data)
IDGFres <- computeIDGF(IDGFdata)
IDGFresrad <- radarIDGF(IDGFres)
head(IDGFresrad)
}
