% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runIDGFTool.R
\name{runIDGFTool}
\alias{runIDGFTool}
\title{Applicatif simplifié IDGF}
\usage{
runIDGFTool()
}
\description{
Cette fonction exécute un applicatif développé via le package Rshiny pour proposer une interface graphique.
}
\examples{
library(IDGF)
# runIDGFTool()
}
