% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taxa.GF}
\alias{taxa.GF}
\title{Jeu de données fictif pour l'IDGF}
\format{
Un tableau avec 575 lignes et 4 variables :
\describe{
  \item{cd_opecont}{l'identifiant du relevé}
  \item{cd_taxon}{le code OMNIDIA du taxon}
  \item{abondance}{abondance relevée du taxon sur 400}
  \item{her}{HER correspondant au relevé}
}
}
\usage{
taxa.GF
}
\description{
Une compilation d'inventaires diatomiques fictifs avec des taxons Guyanais et avec introduction d'un taxon inconnu, "ZZZZ".
}
\keyword{datasets}
