% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeIDGF.R
\name{computeIDGF}
\alias{computeIDGF}
\title{Calcul de l'IDGF}
\usage{
computeIDGF(IDGFdata)
}
\arguments{
\item{IDGFdata}{Données importées par `importIDGF()`}
}
\value{
Renvoie un tableau présentant l'ensemble des résultats (score par métrique, évaluation de l'état écologique, statistiques sur les taxons indiciels/halins/inconnus)
}
\description{
Cette fonction procède au calcul des différentes métriques constitutives de l'IDGF puis à l'évaluation de l'état écologique en se basant sur différentes tables de référence internes et sur l'indication de l'HER dans les données d'entrée.
}
\examples{
library(IDGF)
data <- system.file("input_test.xlsx", package = "IDGF")
IDGFdata <- importIDGF(data)
IDGFres <- computeIDGF(IDGFdata)
head(IDGFres)
}
