% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radar_IDGF.R
\name{Diagnostic_IDGF}
\alias{Diagnostic_IDGF}
\title{Sortie graphique de l'IDGF}
\usage{
Diagnostic_IDGF(result_IDGF, lang = "FR")
}
\arguments{
\item{resultat_IDGF}{Les résultats issus de la fonction IDGF()}
}
\value{
Graphiques radar de chaque relevé présent dans le tableau initial
}
\description{
Calcule l'indice IDGF à partir d'une ou plusieurs listes floristiques
}
\examples{
res.IDGF <- IDGF(taxa.GF, lang = "FR")
Radar_IDGF(res.IDGF)
}
