% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDGF.R
\name{IDGF}
\alias{IDGF}
\title{Indice Diatomique de Guyane Française}
\usage{
IDGF(df, lang = "FR")
}
\arguments{
\item{df}{Le tableau contenant la liste floristique, composé de 4 colonnes (dans cet ordre) : id_releve, cd_taxon, abondance, her (1 ou 2)}

\item{lang}{Argument acceptant deux valeurs : "FR" pour obtenir des résultats en français, "ENG" pour obtenir des résultats en anglais}
}
\value{
Une liste
}
\description{
Calcule l'indice IDGF à partir d'une ou plusieurs listes floristiques
}
\examples{
taxa.GF
IDGF(taxa, lang = "FR")

}
