% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importIDGF.R
\name{importIDGF}
\alias{importIDGF}
\title{Import des données}
\usage{
importIDGF(input)
}
\arguments{
\item{input}{Le nom du fichier de données à importer (ou un chemin)}
}
\value{
une liste
}
\description{
Import des données
}
