% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radarIDGF.R
\name{radarIDGF}
\alias{radarIDGF}
\title{Production de diagrammes de diagnostic issus de l'IDGF}
\usage{
radarIDGF(IDGFres)
}
\arguments{
\item{IDGFres}{sortie de la fonction `computeIDGF()`}
}
\value{
une liste
}
\description{
Production de diagrammes de diagnostic issus de l'IDGF
}
